Rename metadata directory and update imports
